fix ios warning

+++ src/ios/CBPeripheral+Extensions.m
@@ -54,7 +54,7 @@ static char ADVERTISEMENT_RSSI_IDENTIFER;
         id manufacturerData = [advertisementData objectForKey:CBAdvertisementDataManufacturerDataKey];
         if (manufacturerData) {
             const uint8_t *bytes = [manufacturerData bytes];
-            int len = [manufacturerData length];
+            long len = [manufacturerData length];
             // skip manufacturer uuid
             NSData *data = [NSData dataWithBytes:bytes+2 length:len-2];
             [self setAdvertising: [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding]];
